requirements.txt
ConfigArgParse==1.3
javalang==0.13.0
numpy==1.20.1
pandas==1.2.2
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
